+++ web/frontend/src/pages/Projects/index.tsx
@@ -19,7 +19,6 @@ import {
   DeleteOutlined,
   FolderOutlined,
   SearchOutlined,
-  EyeOutlined,
 } from '@ant-design/icons'
 import type { ColumnsType } from 'antd/es/table'
 import dayjs from 'dayjs'